break down server module into component parts - board, player, lobby
